Redirecting to keys generation when no keys are configured
